+++ ClientExample/bin/Debug/log.txt
@@ -80,3 +80,11 @@
 20190121161921 20190121161921 engineers: 2 Id:100002 Name:fabao
 20190121165222 20190121165222 engineers: 2 Id:100001 Name:tumo
 20190121165222 20190121165222 engineers: 2 Id:100002 Name:fabao
+20190122001630 20190122001630 engineers: 2 Id:100001 Name:tumo
+20190122001630 20190122001630 engineers: 2 Id:100002 Name:fabao
+20190122003317 20190122003317 engineers: 2 Id:100001 Name:tumo
+20190122003317 20190122003317 engineers: 2 Id:100002 Name:fabao
+20190122003738 20190122003738 engineers: 2 Id:100001 Name:tumo
+20190122003738 20190122003738 engineers: 2 Id:100002 Name:fabao
+20190122003826 20190122003826 engineers: 2 Id:100001 Name:tumo
+20190122003826 20190122003826 engineers: 2 Id:100002 Name:fabao
