--- Tumo/TmSockets/Components/TmRecvQueue.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Tumo
-{
-    public class TmRecvQueue : TmComponent
-    {
-        public Queue<TmParameter> RecvParameters { get; set; } = new Queue<TmParameter>();
-    }
-}
