--- Tumo/TmSockets/TmTcp/TmSession.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Tumo
-{
-    public class TmSession : TmComponent
-    {
-        
-    }
-}
