+++ Tumo/Tumo.csproj
@@ -80,19 +80,7 @@
     <Compile Include="TmModels\TmSoulers\TmSouler.cs" />
     <Compile Include="TmModels\TmSoulers\TmSoulerItem.cs" />
     <Compile Include="TmModels\TmSoulers\TmSoulerDB.cs" />
-    <Compile Include="TmSockets\Components\TmClient.cs" />
-    <Compile Include="TmSockets\Components\TmServer.cs" />
-    <Compile Include="TmSockets\Components\TmNetClient.cs" />
-    <Compile Include="TmSockets\Components\TmNetServer.cs" />
-    <Compile Include="TmSockets\Components\TmSendQueue.cs" />
-    <Compile Include="TmSockets\Components\TmRecvQueue.cs" />
-    <Compile Include="TmSockets\Components\TmTcp.cs" />
-    <Compile Include="TmSockets\Components\TmQueue.cs" />
-    <Compile Include="TmSockets\Components\TmSocket.cs" />
-    <Compile Include="TmSockets\Systems\TmNetClientSystem.cs" />
-    <Compile Include="TmSockets\Systems\TmNetServerSystem.cs" />
-    <Compile Include="TmSockets\TmTcp\TmNetTcp.cs" />
-    <Compile Include="TmSockets\TmTcp\TmSession.cs" />
+    <Compile Include="TmSockets\Components\TmNetTcp.cs" />
     <Compile Include="TmEntities\TmEcsBases\TmDictionary.cs" />
     <Compile Include="TmEntities\TmHelpers\TmStatic\TmJson.cs" />
     <Compile Include="TmEntities\TmEcsBases\TmSystem.cs" />
@@ -101,15 +89,15 @@
     <Compile Include="TmEntities\TmEcsBases\TmIdGenerater.cs" />
     <Compile Include="TmModels\TmSoulers\TmTransform.cs" />
     <Compile Include="TmModels\TmSoulers\TmUser.cs" />
-    <Compile Include="TmSockets\TmTcp\TClient.cs" />
-    <Compile Include="TmSockets\TmTcp\TmPeer.cs" />
+    <Compile Include="TmSockets\Components\TmClient.cs" />
+    <Compile Include="TmSockets\Components\TmPeer.cs" />
     <Compile Include="TmEntities\TmHelpers\TmStatic\TmResponse.cs" />
-    <Compile Include="TmSockets\TmTcp\TmTcpSession.cs" />
+    <Compile Include="TmSockets\Components\TmTcpSession.cs" />
     <Compile Include="TmEntities\TmHelpers\Components\TmDictionaryParameter.cs" />
     <Compile Include="TmModels\TmEnums\TmInfoType.cs" />
     <Compile Include="TmEntities\TmHelpers\TmStatic\TmLog.cs" />
-    <Compile Include="TmSockets\TmTcp\TmTcpClient.cs" />
-    <Compile Include="TmSockets\TmTcp\TmTcpServer.cs" />
+    <Compile Include="TmSockets\Components\TmTcpClient.cs" />
+    <Compile Include="TmSockets\Components\TmTcpServer.cs" />
     <Compile Include="TmEntities\TmHelpers\TmStatic\TmTimerTool.cs" />
     <Compile Include="TmEntities\TmHelpers\TmStatic\TmMD5Tool.cs" />
     <Compile Include="TmEntities\TmHelpers\Components\TmParameter.cs" />
